+++ main.go
@@ -2,9 +2,10 @@ package main
 
 import (
 	"fmt"
-	"github.com/taskcluster/taskcluster-lib-artifact-go/hashfile"
+	"github.com/taskcluster/taskcluster-lib-artifact-go/prepare"
 )
 
 func main() {
-	fmt.Printf("%+v\n", hashfile.HashFileParts("borap.mp3", 1024, 1024))
+	fmt.Printf("%+v\n", prepare.HashFileParts("borap.mp3", 1024, 1024))
+	fmt.Printf("%+v\n", prepare.HashFile("borap.mp3", 1024))
 }
